+++ src/meta/not_found.hpp
@@ -1,0 +1,12 @@
+// Copyright Steinwurf ApS 2015.
+// All Rights Reserved
+//
+// Distributed under the "BSD License". See the accompanying LICENSE.rst file.
+
+#pragma once
+
+namespace meta
+{
+    /// Represents situations where we did not find a specific type
+    struct not_found {};
+}