+++ src/meta/typelist_fwd.hpp
@@ -1,0 +1,12 @@
+// Copyright Steinwurf ApS 2015.
+// All Rights Reserved
+//
+// Distributed under the "BSD License". See the accompanying LICENSE.rst file.
+
+#pragma once
+
+namespace meta
+{
+    template<class... Types>
+    struct typelist;
+}